feat: change Card and CardList components to fit data from API
