+++ src/types/CardData.tsx
@@ -1,0 +1,20 @@
+export type CardData = {
+  id: number;
+  name: string;
+  status: string;
+  species: string;
+  type: string;
+  gender: string;
+  origin: {
+    name: string;
+    url: string;
+  };
+  location: {
+    name: string;
+    url: string;
+  };
+  image: string;
+  episodes: string[];
+  url: string;
+  created: string;
+};
